# static/swagger.json
{
    "swagger": "2.0",
    "basePath": "/api",
    "paths": {
        "/organize/trip/{trip_id}": {
            "parameters": [
                {
                    "in": "path",
                    "description": "Trip Identifier",
                    "name": "trip_id",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Complete%20User"
                            }
                        }
                    }
                },
                "summary": "Get a list of all organizers for a trip",
                "operationId": "Get a list of all organizers for a trip",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "organize"
                ]
            }
        },
...
        "/registration/trip/{trip_id}": {
            "parameters": [
                {
                    "in": "path",
                    "description": "Trip Identifier",
                    "name": "trip_id",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Complete%20Trip%20User"
                            }
                        }
                    }
                },
                "summary": "Get a list of all users registered for a specific trip",
                "operationId": "Get a list of all users registered for a specific trip",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "registration"
                ]
            }
        },
...
        "/trip/": {
            "post": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Trip%20ID"
                        }
                    }
                },
                "summary": "Add a new trip to the database",
                "operationId": "Create a new trip",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Trip%20without%20ID"
                        }
                    },
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "trip"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Complete%20Trip"
                            }
                        }
                    }
                },
                "summary": "Get a list of all trips in the database",
                "operationId": "List all trips",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "tags": [
                    "trip"
                ]
            }
        },
        "/trip/fromTemplate/{trip_template_id}": {
            "parameters": [
                {
                    "in": "path",
                    "description": "Template Identifier",
                    "name": "trip_template_id",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Complete%20Trip"
                            }
                        }
                    }
                },
                "summary": "Get a list of all trips based on a specific template",
                "operationId": "Get a list of all trips from a specific template",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "trip"
                ]
            }
        },
        "/trip/noTemplate": {
            "post": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Trip%20ID"
                        }
                    }
                },
                "summary": "Add a new trip and a new template (template not specified) to the database",
                "operationId": "Create a new trip and template",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Complete%20Trip%20without%20ID"
                        }
                    },
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "trip"
                ]
            }
        },
        "/trip/{trip_id}": {
            "parameters": [
                {
                    "in": "path",
                    "description": "Trip Identifier",
                    "name": "trip_id",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Complete%20Trip"
                        }
                    }
                },
                "summary": "Get the specified trip",
                "operationId": "Get a specific trip",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "trip"
                ]
            },
            "delete": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Delete the specified trip",
                "operationId": "Delete the specified trip",
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "trip"
                ]
            },
            "put": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Modify the specified trip",
                "operationId": "Modify the specified trip",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Trip%20without%20ID"
                        }
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "trip"
                ]
            }
        },
...
    },
    "info": {
        "title": "MountainMates API",
        "version": "0.1",
        "description": "This API facilitates hiking trips, participant registration, and trip organization."
    },
    "produces": [
        "application/json"
    ],
    "consumes": [
        "application/json"
    ],
    "securityDefinitions": {
        "Bearer Auth": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    },
...
}